requirements.txt
gooddata-sdk~=1.3.0
auth0-python~=4.3.0
pandas~=2.0.3
